fix: correção de códigos e alteração de tipo de dado para amount

+++ src/main/java/com/example/pismo_challenge/DTO/Transaction/Request/RegisterTransactionRequest.java
@@ -3,6 +3,7 @@ package com.example.pismo_challenge.DTO.Transaction.Request;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -12,5 +13,5 @@ public class RegisterTransactionRequest {
     @NotNull
     public int operationTypeCode;
     @NotNull
-    public float amount;
+    public BigDecimal amount;
 }
