+++ src/main/java/com/example/pismo_challenge/Entity/Event/TransactionCreatedEvent.java
@@ -1,5 +1,6 @@
 package com.example.pismo_challenge.Entity.Event;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record TransactionCreatedEvent(UUID accountId, float amount) { }
+public record TransactionCreatedEvent(UUID accountId, BigDecimal amount) { }
